Move RefinanceTokenModal to separate folder
