+++ src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/TableCells/RefinanceTokenModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RefinanceTokenModal'
